fix: fix some bugs

+++ evaluation/align_msg.py
@@ -1,10 +1,12 @@
-import json, sys
+import json, sys, os
 log_dir = sys.argv[1]  # 'logs/Meta-Llama-3.1-70B-Instruct-GPTQ-INT4/'
 MT_0 = json.load(open(log_dir+'HammerBench_Based/MT_0.json','r'))
 MT_res = json.load(open(log_dir+'HammerBench_Based/MT_res.json','r'))
-LLM_check_L = json.load(open(log_dir+'HammerBench_Based/LLM_check_L.json','r'))
-LLM_check_response = json.load(open(log_dir+'HammerBench_Based/LLM_check_response.json','r'))
-turn_flags_res = json.load(open(log_dir+'HammerBench_Based/turn_flags_res.json','r'))
+
+if os.path.exists(log_dir+'HammerBench_Based/LLM_check_L.json'):
+    LLM_check_L = json.load(open(log_dir+'HammerBench_Based/LLM_check_L.json','r'))
+    LLM_check_response = json.load(open(log_dir+'HammerBench_Based/LLM_check_response.json','r'))
+    turn_flags_res = json.load(open(log_dir+'HammerBench_Based/turn_flags_res.json','r'))
 
 test_type = sys.argv[2]  # 'xQxA'
 MT_id = json.load(open(f'data/en/multi-turn/{test_type}.json','r'))
@@ -16,9 +18,10 @@ new_L, new_response, new_turn_flags_res = [], [], []
 for i in MT_id:
     new_0.append(MT_0[i])
     new_res.append(MT_res[i])
-    new_L.append(LLM_check_L[i])
-    new_response.append(LLM_check_response[i])
-    new_turn_flags_res.append(turn_flags_res[i])
+    if os.path.exists(log_dir+'HammerBench_Based/LLM_check_L.json'):
+        new_L.append(LLM_check_L[i])
+        new_response.append(LLM_check_response[i])
+        new_turn_flags_res.append(turn_flags_res[i])
 import os
 if not os.path.exists(log_dir+test_type+'/MT0'):
     os.mkdir(log_dir+test_type+'/MT0')
@@ -28,11 +31,12 @@ with open(log_dir+test_type+'/MT0/MT_0.json', 'w') as fp:
 with open(log_dir+test_type+'/MT0/MT_res.json', 'w') as fp:
     json.dump(new_res ,fp ,indent=2, ensure_ascii=False)
 
-with open(log_dir+test_type+'/MT0/LLM_check_L.json', 'w') as fp:
-    json.dump(new_L ,fp ,indent=2, ensure_ascii=False)
+if os.path.exists(log_dir+'HammerBench_Based/LLM_check_L.json'):
+    with open(log_dir+test_type+'/MT0/LLM_check_L.json', 'w') as fp:
+        json.dump(new_L ,fp ,indent=2, ensure_ascii=False)
 
-with open(log_dir+test_type+'/MT0/LLM_check_response.json', 'w') as fp:
-    json.dump(new_response ,fp ,indent=2, ensure_ascii=False)
+    with open(log_dir+test_type+'/MT0/LLM_check_response.json', 'w') as fp:
+        json.dump(new_response ,fp ,indent=2, ensure_ascii=False)
 
-with open(log_dir+test_type+'/MT0/turn_flags_res.json', 'w') as fp:
-    json.dump(new_turn_flags_res,fp ,indent=2, ensure_ascii=False)
+    with open(log_dir+test_type+'/MT0/turn_flags_res.json', 'w') as fp:
+        json.dump(new_turn_flags_res,fp ,indent=2, ensure_ascii=False)
